Add online results view (#482)
* adjust firebase documentation

* remove average score from history list

* add tabs support to history screen

* remove auth requirement from fetch

* add online results

* refactor

* properly handle missing firebase access

* properly use l10n

* update result spec

* add model name to online screen

* mark invalid results on the result list

* remove unused l10n entry

* fix linter

* enable lan access to firebase functions emulator

* add enable selection button to result list screen

* make online tab a simple class instead of widget

* make history tab a simple class instead of widget

* fix hostory selection mode

* refactor

* refactor: extract tabinterface

* refactor

* refactor

+++ flutter/lib/benchmark/state.dart
@@ -549,7 +549,7 @@ class BenchmarkState extends ChangeNotifier {
                 startDatetime: accuracy.startTime,
                 loadgenValidity: accuracy.validity,
               ),
-        minDurationMs: benchmark.taskConfig.minDurationMs.toDouble(),
+        minDurationMs: benchmark.taskConfig.minDurationMs,
         minSamples: benchmark.taskConfig.minQueryCount,
         backendInfo: BackendReportedInfo(
           filename: backendInfo.libPath,
